+++ src/helpers.ts
@@ -4,19 +4,7 @@ import * as path from 'path';
 import { wordErrorRate as calculateWER } from 'word-error-rate';
 
 import { TranscriptionAnalysisService } from './TranscriptionAnalysisService';
-
-export interface ITestDatum {
-    recording: string;
-    transcription: string;
-}
-
-export type TestData = ITestDatum[];
-
-export interface ITestResult {
-    actualTranscription: string;
-    expectedTranscription: string;
-    wordErrorRate: number;
-};
+import { ITestResult, TestData } from './types';
 
 const validateFile = (filepath: string) => {
     if (filepath === undefined) {
@@ -87,8 +75,7 @@ export const writeToTextFile = (filePath: string, data: Object | Array<Object>) 
  */
 export const handleResponse = (expectedTranscription: string, response: DetailedSpeechPhrase) => {
     try {
-        // const actualTranscription = response.NBest[0].Lexical.toLowerCase();
-        const actualTranscription = `** ()& he_re`;
+        const actualTranscription = response.NBest[0].Lexical.toLowerCase();
 
         const analyzer = new TranscriptionAnalysisService();
