+++ src/types.ts
@@ -1,3 +1,26 @@
+///////////////////////////////////////////////////////////////////////////////
+//
+// helpers
+//
+///////////////////////////////////////////////////////////////////////////////
+export interface ITestDatum {
+    recording: string;
+    transcription: string;
+}
+
+export type TestData = ITestDatum[];
+
+export interface ITestResult {
+    actualTranscription: string;
+    expectedTranscription: string;
+    wordErrorRate: number;
+};
+
+///////////////////////////////////////////////////////////////////////////////
+//
+// TranscriptionAnalysisService
+//
+///////////////////////////////////////////////////////////////////////////////
 export interface UnhandledCharacter {
     character: string;
     sources: UnhandledWord[];
